+++ circuit-foundation/src/commonMain/kotlin/com/slack/circuit/foundation/CircuitContent.kt
@@ -17,8 +17,6 @@ import com.slack.circuit.runtime.presenter.Presenter
 import com.slack.circuit.runtime.screen.PopResult
 import com.slack.circuit.runtime.screen.Screen
 import com.slack.circuit.runtime.ui.Ui
-import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.persistentListOf
 
 @Composable
 public fun CircuitContent(
@@ -54,9 +52,9 @@ public fun CircuitContent(
           newRoot: Screen,
           saveState: Boolean,
           restoreState: Boolean,
-        ): ImmutableList<Screen> {
+        ): List<Screen> {
           onNavEvent(NavEvent.ResetRoot(newRoot, saveState, restoreState))
-          return persistentListOf()
+          return emptyList()
         }
 
         override fun pop(result: PopResult?): Screen? {
@@ -66,7 +64,7 @@ public fun CircuitContent(
 
         override fun peek(): Screen = screen
 
-        override fun peekBackStack(): ImmutableList<Screen> = persistentListOf(screen)
+        override fun peekBackStack(): List<Screen> = listOf(screen)
       }
     }
   CircuitContent(screen, navigator, modifier, circuit, unavailableContent, key)