+++ circuitx/gesture-navigation/src/commonMain/kotlin/com/slack/circuitx/gesturenavigation/GestureNavTransitionHolder.kt
@@ -5,14 +5,13 @@ package com.slack.circuitx.gesturenavigation
 import androidx.compose.runtime.Immutable
 import com.slack.circuit.backstack.NavArgument
 import com.slack.circuit.foundation.animation.AnimatedNavState
-import kotlinx.collections.immutable.ImmutableList
 
 /**
  * A holder class used by the `AnimatedContent` composables. This enables us to pass through all of
  * the necessary information as an argument, which is optimal for `AnimatedContent`.
  */
 @Immutable
-internal data class GestureNavTransitionHolder<T : NavArgument>(val args: ImmutableList<T>) :
+internal data class GestureNavTransitionHolder<T : NavArgument>(val args: List<T>) :
   AnimatedNavState {
-  override val backStack: ImmutableList<NavArgument> = args
+  override val backStack: List<NavArgument> = args
 }