+++ samples/counter/apps/src/jvmMain/kotlin/com/slack/circuit/sample/counter/desktop/DesktopCounterCircuit.kt
@@ -46,7 +46,6 @@ import com.slack.circuit.sample.counter.CounterScreen
 import com.slack.circuit.sample.counter.PrimeScreen
 import com.slack.circuit.sample.counter.Remove
 import com.slack.circuit.sample.counter.buildCircuit
-import kotlinx.collections.immutable.persistentListOf
 
 data object DesktopCounterScreen : CounterScreen
 
@@ -149,7 +148,7 @@ fun main() = application {
     state = WindowState(width = 300.dp, height = 300.dp),
     onCloseRequest = ::exitApplication,
   ) {
-    val initialBackStack = persistentListOf<Screen>(DesktopCounterScreen)
+    val initialBackStack = listOf<Screen>(DesktopCounterScreen)
     val backStack = rememberSaveableBackStack(initialBackStack)
     val navigator = rememberCircuitNavigator(backStack) { exitApplication() }
     val circuit = remember { buildCircuit(uiFactory = CounterUiFactory()) }