+++ samples/star/src/commonMain/kotlin/com/slack/circuit/star/petlist/Filters.common.kt
@@ -7,13 +7,11 @@ import com.slack.circuit.internal.runtime.Parcelize
 import com.slack.circuit.overlay.OverlayHost
 import com.slack.circuit.star.db.Gender
 import com.slack.circuit.star.db.Size
-import kotlinx.collections.immutable.ImmutableSet
-import kotlinx.collections.immutable.toImmutableSet
 
 @Parcelize
 class Filters(
-  val genders: ImmutableSet<Gender> = Gender.entries.asIterable().toImmutableSet(),
-  val sizes: ImmutableSet<Size> = Size.entries.asIterable().toImmutableSet(),
+  val genders: Set<Gender> = Gender.entries.toSet(),
+  val sizes: Set<Size> = Size.entries.toSet(),
 ) : Parcelable
 
 /** Indirection for showing filters in the pet list screen. */